Consolidate all actions for DA
