+++ websites/recipe-website/editor/controller/actions/index.ts
@@ -25,6 +25,22 @@ import {
   writeRecipeToIndex,
 } from "./common";
 
+import simpleGit, { SimpleGit } from "simple-git";
+import { getContentDirectory } from "content-engine/fs/getContentDirectory";
+import { directoryIsGitRepo } from "content-engine/git/commit";
+
+import { z } from "zod";
+
+import { writeFile } from "fs-extra";
+import { join } from "node:path";
+
+const INITIAL_COMMIT_MESSAGE = "Initial commit";
+
+const remoteSchema = z.object({
+  remoteName: z.string().min(1, "Remote Name is required"),
+  remoteUrl: z.string().min(1, "Remote URL is required"),
+});
+
 export async function rebuildRecipeIndex() {
   const db = getRecipeDatabase();
   await db.drop();
@@ -234,4 +250,163 @@ export async function deleteRecipe(date: number, slug: string) {
   revalidatePath("/recipe/" + slug);
   revalidatePath("/");
   redirect("/");
+}
+
+export async function createRemote(
+  _state: string | undefined,
+  formData: FormData,
+) {
+  const contentDirectory = getContentDirectory();
+  const result = remoteSchema.safeParse({
+    remoteName: formData.get("remoteName"),
+    remoteUrl: formData.get("remoteUrl"),
+  });
+
+  if (!result.success) {
+    return (
+      result.error.flatten().fieldErrors.remoteName?.[0] ??
+      result.error.flatten().fieldErrors.remoteUrl?.[0]
+    );
+  }
+
+  if (await directoryIsGitRepo(contentDirectory)) {
+    try {
+      const git = simpleGit({
+        baseDir: contentDirectory,
+      });
+      await git.addRemote(result.data.remoteName, result.data.remoteUrl);
+    } catch (e) {
+      if (
+        e &&
+        typeof e === "object" &&
+        "message" in e &&
+        typeof e.message === "string"
+      ) {
+        return e.message;
+      } else {
+        throw e;
+      }
+    }
+  }
+  revalidatePath("/git");
+}
+
+export async function createBranch(
+  _state: string | undefined,
+  formData: FormData,
+) {
+  const contentDirectory = getContentDirectory();
+  const branchName = formData.get("branchName") as string;
+  if (!branchName) {
+    return "Branch Name is required";
+  }
+  if (await directoryIsGitRepo(contentDirectory)) {
+    try {
+      await simpleGit(contentDirectory).checkout(["-b", branchName]);
+    } catch (e) {
+      if (
+        e &&
+        typeof e === "object" &&
+        "message" in e &&
+        typeof e.message === "string"
+      ) {
+        return e.message;
+      } else {
+        throw e;
+      }
+    }
+  }
+  revalidatePath("/git");
+}
+
+const commandHandlers: Record<
+  string,
+  (args: { git: SimpleGit; branch: string }) => Promise<void>
+> = {
+  async checkout({ git, branch }) {
+    if (!branch) {
+      throw new Error("Invalid branch");
+    }
+    await git.checkout(branch);
+    await rebuildRecipeIndex();
+  },
+  async delete({ git, branch }) {
+    if (!branch) {
+      throw new Error("Invalid branch");
+    }
+    await git.deleteLocalBranch(branch);
+  },
+  async forceDelete({ git, branch }) {
+    if (!branch) {
+      throw new Error("Invalid branch");
+    }
+    await git.deleteLocalBranch(branch, true);
+  },
+};
+
+export async function branchCommandAction(
+  _previousState: string | null,
+  formData: FormData,
+): Promise<string | null> {
+  const command = formData.get("command");
+  if (typeof command !== "string") {
+    return "No command provided!";
+  }
+  const commandHandler = commandHandlers[command];
+  if (!commandHandler) {
+    return `Invalid command: ${command}`;
+  }
+  const branch = formData.get("branch");
+  if (typeof branch !== "string") {
+    return `Invalid branch`;
+  }
+  const contentDirectory = getContentDirectory();
+  if (!(await directoryIsGitRepo(contentDirectory))) {
+    return "Content directory is not a Git repository.";
+  }
+  try {
+    const git = simpleGit({
+      baseDir: contentDirectory,
+    });
+    await commandHandler({ git, branch });
+  } catch (e) {
+    if (
+      e &&
+      typeof e === "object" &&
+      "message" in e &&
+      typeof e.message === "string"
+    ) {
+      return e.message;
+    } else {
+      throw e;
+    }
+  }
+  await rebuildRecipeIndex();
+  revalidatePath("/git");
+  return null;
+}
+
+export async function remoteCommandAction(
+  _previousState: string | null,
+  _formData: FormData,
+): Promise<string | null> {
+  return null;
+}
+
+export async function initializeContentGit() {
+  const contentDirectory = getContentDirectory();
+  if (!(await directoryIsGitRepo(contentDirectory))) {
+    const git = simpleGit({
+      baseDir: contentDirectory,
+    });
+    await git.init();
+    await writeFile(
+      join(contentDirectory, ".gitignore"),
+      `/transformed-images
+/recipes/index`,
+    );
+    await git.add(".");
+    await git.commit(INITIAL_COMMIT_MESSAGE);
+  }
+  revalidatePath("/git");
 }
