--- websites/recipe-website/editor/src/actions/deleteRecipe/index.ts
@@ -1,15 +1,0 @@
-"use server";
-
-import { auth, signIn } from "@/auth";
-import { deleteRecipe } from "recipe-editor/controller/actions";
-
-export default async function authenticateAndDeleteRecipe(
-  currentDate: number,
-  currentSlug: string,
-) {
-  const user = await auth();
-  if (!user) {
-    return signIn();
-  }
-  return deleteRecipe(currentDate, currentSlug);
-}
