+++ websites/recipe-website/editor/src/app/(editor)/git/BranchSelector.tsx
@@ -2,7 +2,7 @@
 
 import { useActionState, useState } from "react";
 import { SubmitButton } from "component-library/components/SubmitButton";
-import { branchCommandAction } from "./actions";
+import { branchCommandAction } from "../../../../controller/actions";
 import clsx from "clsx";
 import { BranchSummaryBranch } from "simple-git";
